rename to _distrib_checks
